+++ packages/dso-toolkit/src/components/anchor/index.ts
@@ -1,3 +1,0 @@
-export * from "./anchor.args.js";
-export * from "./anchor.models.js";
-export * from "./anchor.stories-of.js";