+++ packages/dso-toolkit/src/components/contact-information/contact-information.models.ts
@@ -1,8 +1,8 @@
-import { Anchor } from "../anchor";
+import { Link } from "../link";
 import { Heading } from "../heading";
 
 export interface ContactInformation<TemplateFnReturnType> {
   heading?: Heading<TemplateFnReturnType>;
-  anchorItems?: Anchor[];
+  anchorItems?: Link[];
   infoItems?: string[];
 }