+++ storybook/src/components/contact-information/contact-information.core-template.ts
@@ -6,7 +6,7 @@ import { ContactInformation } from "dso-toolkit";
 export const coreContactInformation: ComponentImplementation<ContactInformation<TemplateResult>> = {
   component: "contactInformation",
   implementation: "core",
-  template: ({ anchorTemplate, headingTemplate }) =>
+  template: ({ linkTemplate, headingTemplate }) =>
     function contactInformationTemplate({ heading, anchorItems, infoItems }) {
       return html`<dso-contact-information>
         ${heading
@@ -18,7 +18,7 @@ export const coreContactInformation: ComponentImplementation<ContactInformation<
         ${anchorItems && anchorItems.length > 0
           ? html`
               <ul slot="anchor-items">
-                ${anchorItems.map((anchor) => html`<li>${anchorTemplate(anchor)}</li>`)}
+                ${anchorItems.map((anchor) => html`<li>${linkTemplate(anchor)}</li>`)}
               </ul>
             `
           : nothing}