+++ storybook/src/components/link/link.css-template.ts
@@ -1,13 +1,13 @@
-import { Anchor } from "dso-toolkit";
+import { Link } from "dso-toolkit";
 import { html, nothing } from "lit-html";
 import { ifDefined } from "lit-html/directives/if-defined.js";
 import { ComponentImplementation } from "../../templates";
 
-export const cssAnchor: ComponentImplementation<Anchor> = {
-  component: "anchor",
+export const cssLink: ComponentImplementation<Link> = {
+  component: "link",
   implementation: "html-css",
   template: ({ iconTemplate }) =>
-    function anchorTemplate({ icon, iconMode, label, modifier, mode, url, ariaCurrent }) {
+    function linkTemplate({ icon, iconMode, label, modifier, mode, url, ariaCurrent }) {
       return html`<a
         href=${url}
         class=${ifDefined([modifier, mode].filter((c) => !!c).join(" ") || undefined)}