+++ storybook/src/components/onboarding-tip/onboarding-tip.content.ts
@@ -2,12 +2,12 @@ import { html } from "lit-html";
 import { Templates } from "../../templates";
 
 export function richContent(templates: Templates) {
-  const { anchorTemplate, buttonRowTemplate, imageTemplate, richContentTemplate } = templates;
+  const { linkTemplate, buttonRowTemplate, imageTemplate, richContentTemplate } = templates;
 
   return html`${richContentTemplate({
     children: html`<p>
         Uitleg over wat u hier kunt doen en lezen. Met een link naar
-        ${anchorTemplate({ label: "Meer informatie", url: "#", iconMode: "after", icon: { icon: "external-link" } })}.
+        ${linkTemplate({ label: "Meer informatie", url: "#", iconMode: "after", icon: { icon: "external-link" } })}.
       </p>
       <p>
         En ook een afbeelding: ${imageTemplate({ source: "images/sneeuwpop.png", alt: "Afbeelding van een sneeuwpop" })}